Switched to Qt integer types, due to the problems with QVariant on Qt < 5.6

+++ admin/studentform.cpp
@@ -112,7 +112,7 @@ bool StudentForm::insertRecord(QSqlRecord &record, QSqlError &error) {
     auto success = manager().saveStudent(student, error);
     if (success) {
         refreshModel();
-        record.setValue("id", static_cast<quint64>(student.id()));
+        record.setValue("id", student.id());
     }
     return success;
 }
