+++ common/course.cpp
@@ -9,7 +9,7 @@ namespace paso {
 namespace data {
 namespace entity {
 
-Course::Course(const QString &code, const QString &name, uint64_t id)
+Course::Course(const QString &code, const QString &name, quint64 id)
     : Entity(id), mCode(code), mName(name) {}
 
 Course::Course(const QVariantMap &map)
@@ -75,7 +75,7 @@ QSqlQuery Course::updateQuery(const QSqlDatabase &database,
     QSqlQuery query(database);
     query.prepare(
         "UPDATE COURSE SET CODE = :code, NAME = :name WHERE ID = :id");
-    query.bindValue(":id", static_cast<quint64>(course.id()));
+    query.bindValue(":id", course.id());
     query.bindValue(":code", course.code());
     query.bindValue(":name", course.name());
     return query;
